fix add&update title bug

+++ dlxy-article-server/dlxy-article-server-dao/src/main/java/com/dlxy/server/article/dao/mybatis/DlxyTitleDao.java
@@ -7,7 +7,7 @@ import com.dlxy.common.dto.DlxyTitleDTO;
 import com.dlxy.server.article.model.DlxyTitleExample;
 
 public interface DlxyTitleDao {
-    int countByExample(DlxyTitleExample example);
+    long countByExample(DlxyTitleExample example);
 
     int deleteByExample(DlxyTitleExample example);
